pyproject.toml
[project]
name = "forge-cli"
version = "0.1.0"
description = "AI agent failure mode tracking and analysis — USMI Labs"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39.0"]
openai = ["openai>=1.0.0"]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[project.scripts]
forge = "forge_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forge_cli"]

[tool.ruff]
target-version = "py311"
line-length = 100
